Refactor: set User Package to summoner
