+++ src/main/java/mejai/mejaigg/summoner/dto/response/UserStreakDto.java
@@ -1,4 +1,4 @@
-package mejai.mejaigg.user.dto.response;
+package mejai.mejaigg.summoner.dto.response;
 
 import java.time.format.DateTimeFormatter;
